feat: Remove DocumentsExtensions portlet from bottom-all-container configuration - EXO-70137
Prior to this change, in bottom-all-container, we inject and load documents extensions as a portlet, so, it is loaded on all pages even if no need to be used which can lead to loading performance issues.
To fix that, we must use js load-group, which are now available through all webapps to load documents extensions only with favorite drawer.
