--- documents-webapp/src/main/webapp/html/documentsExtensions.html
@@ -1,3 +1,0 @@
-<script>
-  require(['PORTLET/documents-portlet/DocumentsExtensions'], extensions => extensions.init());
-</script>
