--- documents-webapp/src/main/webapp/vue-app/documents-extensions/initComponents.js
@@ -1,8 +1,0 @@
-import DocumentsFavoriteItem from './components/DocumentsFavoriteItem.vue';
-const components = {
-  'documents-favorite-item': DocumentsFavoriteItem,
-};
-
-for (const key in components) {
-  Vue.component(key, components[key]);
-}
