+++ documents-webapp/src/main/webapp/vue-app/documents-favorite-drawer-extensions/main.js
@@ -14,24 +14,11 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
+ 
+import './initComponents.js';
+import {initDocumentsExtensions} from './extensions.js';
+import '../documents-icons-extension/extensions.js';
 
-export function injectSortTooltip(tooltipName,markerClass){
-  document.getElementsByClassName(markerClass).forEach(element => {
-    element.getElementsByTagName('i').item(0);
-    element.title = tooltipName; 
-  });
-}
-
-export function getSize(size) {
-  if ( size === 0){
-    return  {value: 0, unit: 'B'};
-  } 
-  const m = size > 0 ? 1 : -1;
-  const k = Math.floor((Math.log2(Math.abs(size)) / 10));
-  let rank = `B`;
-  if (k!==0) {
-    rank = `${'KMGT'[k - 1]}B`;
-  }    
-  const count = (Math.abs(size) / Math.pow(1024, k)).toFixed(2);
-  return {value: Math.round(count*m), unit: rank};
+export function init() {
+  initDocumentsExtensions();
 }
