+++ documents-webapp/webpack.prod.js
@@ -36,7 +36,7 @@ const config = {
   entry: {
     documents: './src/main/webapp/vue-app/documents/main.js',
     documentsSnackbarComponent : './src/main/webapp/vue-app/snackbar/main.js',
-    documentsExtensions: './src/main/webapp/vue-app/documents-extensions/main.js',
+    documentsFavoriteDrawerExtensions: './src/main/webapp/vue-app/documents-favorite-drawer-extensions/main.js',
     documentsSizeGadget: './src/main/webapp/vue-app/documents-size-gadget/main.js',
     downloadDocumentsPublicAccess: './src/main/webapp/vue-app/download-document/main.js',
     notificationExtension: './src/main/webapp/vue-app/notification-extension/main.js',
